backend/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.15
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3
